+++ SmallServerAdmin/settings/CHANGELOG.md
@@ -3,8 +3,12 @@
 All notable changes to this module will be documented in this file.
 
 ## [1.0.4] - 16.07.2016
+### Added
+- Setting write permissions to the directory that contains the configuration files of a servers (*/servers*).
+
 ### Fixed
-- Saving settings, which did not exist in the configuration file ([#34](https://github.com/adminstock/ssa/issues/34));.
+- Saving settings, which did not exist in the configuration file ([#34](https://github.com/adminstock/ssa/issues/34));
+- Fixed problem with incorrect output information in the server list (`GetServerInfo` in *api.php*).
 
 ## [1.0.3] - 02.07.2016
 ### Added